This commit renames some of the files and adds new files This keeps the files more maintainable and scalable
