+++ app.js
@@ -2,8 +2,10 @@ const express = require('express')
 const app = express()
 const cors = require('cors');
 const {apiRouter, articlesRouter, topicRouter, commentRouter, userRouter} = require('./routes/index-router')
-const { noEndpoint} = require('./controllers/api.controller')
-const {handleCustomErrors, handlePsqlErrors, handleServerErrors} = require('./errorHandler/api.errorHandler')
+const { noEndpoint } = require('./controllers/api-controllers')
+const { handleCustomErrors } = require('./errorHandler/handleCustomErrors')
+const { handlePsqlErrors } = require('./errorHandler/handlePsqlErrors')
+const { handleServerErrors } = require('./errorHandler/handleServerErrors')
 
 app.use(cors());
