+++ db/connection.js
@@ -12,11 +12,8 @@ if (ENV === 'production') {
   config.max = 2;
 }
 
-
-
 if (!process.env.PGDATABASE && !process.env.DATABASE_URL) {
   throw new Error('PGDATABASE or DATABASE_URL not set');
 }
-
 
 module.exports = new Pool(config);
