+++ errorHandler/handleServerErrors.js
@@ -1,0 +1,3 @@
+exports.handleServerErrors = ((err, req, res, next)=>{
+    res.status(500).send({msg: 'Internal Server Error'})
+})
