+++ routes/api-routes.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const apiRouter = express.Router()
-const { endpointDocumentation } = require('../controllers/api.controller')
+const { endpointDocumentation } = require('../controllers/api-controllers')
 
 apiRouter
         .get('/', endpointDocumentation)
